# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: TariffShield API
  description: API for managing customs-bond collateral and importer/surety auth
  version: 1.0.0
servers:
  - url: http://localhost:3002
paths:
  /auth/signup:
    post:
      summary: Importer or Surety Sign Up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minimum: 8
                role:
                  type: string
                  enum: [importer, surety_admin]
                  default: importer
                privacyPolicyVersionId:
                  type: string
      responses:
        '200':
          description: Successful signup
        '400':
          description: Invalid input
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: User Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successful login
        '400':
          description: Invalid input
        '401':
          description: Invalid credentials
        '429':
          description: Too many attempts
...
  /importers:
    get:
      summary: List registered importers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of importers
    post:
      summary: Register a new importer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - legalName
                - bondId
                - initialRequiredCollateral
              properties:
                legalName:
                  type: string
                ein:
                  type: string
                bondId:
                  type: integer
                initialRequiredCollateral:
                  type: string
      responses:
        '200':
          description: Importer registered
        '403':
          description: Forbidden or blocked by AML/OFAC
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT